requirements-dev.txt
-r requirements.txt
aioresponses==0.7.1
flake8==3.8.4
mypy==0.790
mypy-extensions==0.4.3
psycopg2-binary==2.8.6
pytest==5.4.3
pytest-aiohttp==0.3.0
pytest-asyncio==0.14.0
